requirements.txt
# Voice Sales Agent MVP - Requirements

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# OpenAI Agents SDK with voice support
openai-agents[voice]>=0.1.0

# OpenAI SDK
openai>=1.0.0

# Pinecone - Vector database for RAG
pinecone>=3.0.0

# Data validation
pydantic>=2.0.0

# Audio processing
numpy>=1.24.0
sounddevice>=0.4.6

# Utilities
python-dotenv>=1.0.0

# Async support
aiohttp>=3.9.0

# MongoDB - Async driver
motor>=3.3.0
pymongo>=4.6.0
